app/requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
Django==2.2.13
django-material==1.3.0
django-session-timeout==0.0.3
djangorestframework==3.11.0
idna==2.7
pytz==2020.1
requests==2.19.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.23
